# openapi.yaml
openapi: 3.0.0
info:
  title: API de Usuarios
  version: 1.0.0
paths:
  /users:
    post:
      summary: Registrar un nuevo usuario
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: Usuario elimina correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Correo ya registrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Servicio no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'



    get:
      summary: Consultar múltiples usuarios con paginación
      operationId: getUsers
      parameters:
        - name: page
          in: query
          required: false
          description: Número de la página a recuperar
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          required: false
          description: Cantidad de usuarios por página
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Lista de usuarios recuperada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'

        '401':
            description: No autorizado, token inválido o ausente.
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/ErrorResponse'
        '403':
            description: Prohibido, sin permisos para acceder a los usuarios.
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '500':
            description: Error interno del servidor
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
        '503':
            description: Servicio no disponible
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'




  /users/{id}/password:
    patch:
      summary: Actualizar la contraseña de un usuario
      operationId: updateUserPassword
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordUpdate'
      responses:
        '200':
          description: Contraseña actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Prohibido (sin permisos suficientes)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Servicio no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{id}:
    delete:
      summary: Eliminar un usuario existente
      operationId: deleteUser
      security:
        - BearerAuth: [ ]
      parameters:
          - name: id
            in: path
            required: true
            schema:
              type: string


      responses:
        '200':
          description: Usuario eliminado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteUserResponse'
        '401':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Permisos insuficientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Servicio no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Obtener un usuario
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string


      responses:
          '200':
            description: Usuario obtenido exitosamente
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/UserResponse'
          '401':
            description: Datos de entrada inválidos
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/ErrorResponse'

          '403':
            description: Permisos insuficientes
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/ErrorResponse'
          '404':
            description: Usuario no encontrado
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '500':
            description: Error interno del servidor
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
          '503':
            description: Servicio no disponible
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Actualizar un usuario
      description: Actualiza la información de un usuario existente.
      parameters:
        - name: id
          in: path
          required: true
          description: ID del usuario a actualizar
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'

      responses:
        '200':
          description: Usuario actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          description: Datos de entrada inválidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Permisos insuficientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflicto, datos en uso o inconsistencia
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Servicio no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'







components:
  schemas:
    UserRegistration:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).*$"
          description: La contraseña del usuario (debe contener al menos un dígito, una mayúscula y una minúscula)
        fullName:
          type: string
          maxLength: 100
        dateBirth:
          type: string
          format: date
          example: "1990-01-01"
        rol:
          type: string
          enum:
            - user
            - admin
          default: user
      required:
        - fullName
        - email
        - password
    PasswordUpdate:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          description: Contraseña actual del usuario
        newPassword:
          type: string
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).*$"
          description: Nueva contraseña del usuario (debe contener al menos un dígito, una mayúscula y una minúscula)





    UserGetById:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: Identificador único del usuario
    UserUpdate:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).*$"
          description: La contraseña del usuario (debe contener al menos un dígito, una mayúscula y una minúscula)
        fullName:
          type: string
          maxLength: 100
        dateBirth:
          type: string
          format: date
          example: "1990-01-01"
        rol:
          type: string
          enum:
            - user
            - admin
          default: user
      required:
        - fullName
        - email
        - password
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        pagination:
          type: object
          properties:
            page:
              type: integer
            size:
              type: integer
            totalPages:
              type: integer
            totalItems:
              type: integer

    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "Contraseña actualizada exitosamente"
      required:
        - message
    UserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        dateBirth:
          type: string
          format: date
        rol:
          type: string
      required:
        - id
        - email
        - fullName
        - rol
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    DeleteUserResponse:
      type: object
      properties:
        message:
          type: string
          example: "Usuario eliminado correctamente"
        userId:
          type: integer
          example: 123

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT